converted app pages to regular compoennts, moved client side components to components folder
