+++ src/app/(auth)/signin/page.tsx
@@ -1,122 +1,9 @@
-"use client";
+import SignInForm from "@/components/SigninPage";
 
-import { useState } from "react";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import {
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { SignInFormSchema } from "@/utils/zod/form";
-import useFetch from "@/hooks/useFetch";
-import { useRouter } from "next/navigation";
-import { Database } from "@/types/db";
-
-type User = Database['public']['Tables']['users']['Row']
-export default function SignInForm() {
-  const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
-  const { fetchData } = useFetch();
-
-  const form = useForm<z.infer<typeof SignInFormSchema>>({
-    resolver: zodResolver(SignInFormSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-  });
-
-  function onSubmit(values: z.infer<typeof SignInFormSchema>) {
-    setIsLoading(true);
-    fetchData<User>("/api/auth/signin", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(values),
-    }).then(() => {
-      router.replace("/");
-    }).catch((error) => {
-      console.error(error);
-    }).finally(() => {
-      setIsLoading(false);
-    });
-  }
-
+export default function SignInPage() {
   return (
-    <>
-      <CardHeader>
-        <CardTitle className="text-2xl">Sign In</CardTitle>
-        <CardDescription>
-          Enter your credentials to access your account
-        </CardDescription>
-      </CardHeader>
-      <CardContent>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-            <FormField
-              control={form.control}
-              name="email"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Email</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="email"
-                      placeholder="john.doe@example.com"
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <FormField
-              control={form.control}
-              name="password"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Password</FormLabel>
-                  <FormControl>
-                    <Input type="password" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <Button type="submit" className="w-full" disabled={isLoading}>
-              {isLoading ? "Signing in..." : "Sign In"}
-            </Button>
-          </form>
-        </Form>
-      </CardContent>
-      <CardFooter className="flex justify-center border-t p-4">
-        <p className="text-sm text-muted-foreground">
-          Don&apost have an account?{" "}
-          <a
-            href="/register"
-            className="text-primary underline underline-offset-4"
-          >
-            Register
-          </a>
-        </p>
-      </CardFooter>
-    </>
+    <div>
+      <SignInForm />
+    </div>
   );
 }
